重构控制器

+++ src/main/java/io/github/springsongs/controller/SpringAritlceController.java
@@ -6,7 +6,6 @@ import java.util.List;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import javax.validation.constraints.NotEmpty;
-import javax.validation.constraints.NotNull;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,7 +13,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.util.StringUtils;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -25,12 +24,11 @@ import org.springframework.web.bind.annotation.RestController;
 import io.github.springsongs.domain.dto.ReponseResultPageDTO;
 import io.github.springsongs.domain.dto.ResponseDTO;
 import io.github.springsongs.domain.dto.SpringAritlceDTO;
-import io.github.springsongs.domain.query.SpringAritlceQuery;
+import io.github.springsongs.domain.query.SpringAritlceQueryBO;
 import io.github.springsongs.enumeration.ResultCode;
 import io.github.springsongs.service.ISpringAritlceService;
 import io.github.springsongs.util.Constant;
 import io.github.springsongs.util.IpKit;
-import io.github.springsongs.util.R;
 
 @RestController
 @RequestMapping(value = "/SpringAritlce")
@@ -42,7 +40,7 @@ public class SpringAritlceController extends BaseController {
 	private ISpringAritlceService springAritlceService;
 
 	@PostMapping(value = "/ListByPage")
-	public ReponseResultPageDTO<SpringAritlceDTO> listByPage(@RequestBody SpringAritlceQuery springAritlceQuery,
+	public ReponseResultPageDTO<SpringAritlceDTO> listByPage(@RequestBody SpringAritlceQueryBO springAritlceQuery,
 			@PageableDefault(page = 1, size = 20) Pageable pageable) {
 		Page<SpringAritlceDTO> lists = springAritlceService.getAllRecordByPage(springAritlceQuery, pageable);
 		return ReponseResultPageDTO.successed(lists.getContent(), lists.getTotalElements(),
@@ -76,15 +74,19 @@ public class SpringAritlceController extends BaseController {
 	}
 
 	@PostMapping(value = "/SetDeleted")
-	public ResponseDTO<String> setDeleted(
-			@RequestParam(value = "ids", required = true) @Valid @NotNull(message = Constant.PARAMETER_NOT_NULL_ERROR) List<String> ids) {
-
+	public ResponseDTO<String> setDeleted(@RequestParam(value = "ids") List<String> ids) {
+		if (CollectionUtils.isEmpty(ids)) {
+			return ResponseDTO.successed(null, ResultCode.PARAMETER_NOT_NULL_ERROR);
+		}
 		springAritlceService.setDeleted(ids);
 		return ResponseDTO.successed(null, ResultCode.DELETE_SUCCESSED);
 	}
 
 	@PostMapping(value = "/Deleted")
-	public ResponseDTO<String> deleted(@RequestParam(value = "ids", required = true) List<String> ids) {
+	public ResponseDTO<String> deleted(@RequestParam(value = "ids") List<String> ids) {
+		if (CollectionUtils.isEmpty(ids)) {
+			return ResponseDTO.successed(null, ResultCode.PARAMETER_NOT_NULL_ERROR);
+		}
 		springAritlceService.delete(ids);
 		return ResponseDTO.successed(null, ResultCode.DELETE_SUCCESSED);
 	}
@@ -97,19 +99,19 @@ public class SpringAritlceController extends BaseController {
 	}
 
 	@PostMapping(value = "/HotStatus/{id}")
-	public ResponseDTO<String> hotStatus(@PathVariable(value = "id", required = true) String id) {
+	public ResponseDTO<String> hotStatus(@PathVariable(value = "id", required = true)  @Valid @NotEmpty(message = Constant.PARAMETER_NOT_NULL_ERROR) String id) {
 		springAritlceService.hotStatus(id);
 		return ResponseDTO.successed(null, ResultCode.SAVE_SUCCESSED);
 	}
 
 	@PostMapping(value = "/TopStatus/{id}")
-	public ResponseDTO<String> topStatus(@PathVariable(value = "id", required = true) String id) {
+	public ResponseDTO<String> topStatus(@PathVariable(value = "id", required = true)  @Valid @NotEmpty(message = Constant.PARAMETER_NOT_NULL_ERROR) String id) {
 		springAritlceService.topStatus(id);
 		return ResponseDTO.successed(null, ResultCode.SAVE_SUCCESSED);
 	}
 
 	@PostMapping(value = "/Featured/{id}")
-	public ResponseDTO<String> featured(@PathVariable(value = "id", required = true) String id) {
+	public ResponseDTO<String> featured(@PathVariable(value = "id", required = true)  @Valid @NotEmpty(message = Constant.PARAMETER_NOT_NULL_ERROR) String id) {
 		springAritlceService.featured(id);
 		return ResponseDTO.successed(null, ResultCode.SAVE_SUCCESSED);
 	}
